refactor: arrange compoennts adequately
