+++ packages/reactnative/src/components/modals/ImportAccountModal.tsx
@@ -14,7 +14,7 @@ import globalStyles from '../../styles/globalStyles';
 import { Wallet } from '../../types/wallet';
 import { COLORS } from '../../utils/constants';
 import { FONT_SIZE, WINDOW_WIDTH } from '../../utils/styles';
-import Button from '../Button';
+import Button from '../buttons/CustomButton';
 
 type Props = {
   modal: {
