+++ packages/reactnative/src/components/modals/ImportTokenModal.tsx
@@ -15,8 +15,8 @@ import { addToken, Token } from '../../store/reducers/Tokens';
 import globalStyles from '../../styles/globalStyles';
 import { COLORS } from '../../utils/constants';
 import { FONT_SIZE, WINDOW_WIDTH } from '../../utils/styles';
-import Blockie from '../Blockie';
-import Button from '../Button';
+import Button from '../buttons/CustomButton';
+import Blockie from '../scaffold-eth/Blockie';
 
 type Props = {
   modal: {
