+++ packages/reactnative/src/components/modals/NFTTransferConfirmationModal.tsx
@@ -8,7 +8,7 @@ import { Account } from '../../store/reducers/Accounts';
 import globalStyles from '../../styles/globalStyles';
 import { parseFloat, truncateAddress } from '../../utils/helperFunctions';
 import { FONT_SIZE } from '../../utils/styles';
-import Blockie from '../Blockie';
+import Blockie from '../scaffold-eth/Blockie';
 import Fail from './modules/Fail';
 import Success from './modules/Success';
