+++ packages/reactnative/src/components/modals/PrivateKeyModal.tsx
@@ -14,8 +14,8 @@ import { Wallet } from '../../types/wallet';
 import { COLORS } from '../../utils/constants';
 import { truncateAddress } from '../../utils/helperFunctions';
 import { FONT_SIZE, WINDOW_WIDTH } from '../../utils/styles';
-import Blockie from '../Blockie';
-import Button from '../Button';
+import Button from '../buttons/CustomButton';
+import Blockie from '../scaffold-eth/Blockie';
 
 type Props = {
   modal: {
