+++ packages/reactnative/src/components/modals/SeedPhraseModal.tsx
@@ -10,7 +10,7 @@ import globalStyles from '../../styles/globalStyles';
 import { Security } from '../../types/security';
 import { COLORS } from '../../utils/constants';
 import { FONT_SIZE, WINDOW_WIDTH } from '../../utils/styles';
-import Button from '../Button';
+import Button from '../buttons/CustomButton';
 
 type Props = {
   modal: {
