+++ packages/reactnative/src/components/modals/SignMessageModal.tsx
@@ -7,7 +7,7 @@ import useNetwork from '../../hooks/scaffold-eth/useNetwork';
 import globalStyles from '../../styles/globalStyles';
 import { parseBalance } from '../../utils/helperFunctions';
 import { FONT_SIZE, WINDOW_HEIGHT, WINDOW_WIDTH } from '../../utils/styles';
-import Blockie from '../Blockie';
+import Blockie from '../scaffold-eth/Blockie';
 
 type Props = {
   modal: {
