+++ packages/reactnative/src/components/modals/TxReceiptModal.tsx
@@ -7,7 +7,7 @@ import useNetwork from '../../hooks/scaffold-eth/useNetwork';
 import globalStyles from '../../styles/globalStyles';
 import { COLORS } from '../../utils/constants';
 import { FONT_SIZE, WINDOW_WIDTH } from '../../utils/styles';
-import Button from '../Button';
+import Button from '../buttons/CustomButton';
 
 type Props = {
   modal: {
