+++ packages/reactnative/src/components/modals/modules/Fail.tsx
@@ -3,7 +3,7 @@ import { Dimensions, Image, StyleSheet, View } from 'react-native';
 import { Modal, Portal, Text } from 'react-native-paper';
 import globalStyles from '../../../styles/globalStyles';
 import { FONT_SIZE } from '../../../utils/styles';
-import Button from '../../Button';
+import Button from '../../buttons/CustomButton';
 
 type Props = {
   isVisible: boolean;
