+++ packages/reactnative/src/components/scaffold-eth/Address.tsx
@@ -7,7 +7,7 @@ import globalStyles from '../../styles/globalStyles';
 import { COLORS } from '../../utils/constants';
 import { truncateAddress } from '../../utils/helperFunctions';
 import { FONT_SIZE } from '../../utils/styles';
-import Blockie from '../Blockie';
+import Blockie from './Blockie';
 
 type Props = {
   address: string;
