+++ packages/reactnative/src/screens/Authentication/CreateWallet.tsx
@@ -7,8 +7,8 @@ import { useToast } from 'react-native-toast-notifications';
 import { useDispatch } from 'react-redux';
 import BackButton from '../../components/buttons/BackButton';
 import SeedPhrase from '../../components/SeedPhrase';
+import useWallet from '../../hooks/scaffold-eth/useWallet';
 import { useSecureStorage } from '../../hooks/useSecureStorage';
-import useWallet from '../../hooks/useWallet';
 import { initAccounts } from '../../store/reducers/Accounts';
 import { loginUser } from '../../store/reducers/Auth';
 import globalStyles from '../../styles/globalStyles';
