+++ packages/reactnative/src/screens/Main/Tab/modules/wallet/Header.tsx
@@ -12,8 +12,8 @@ import Share from 'react-native-share';
 import { useToast } from 'react-native-toast-notifications';
 // @ts-ignore
 import Ionicons from 'react-native-vector-icons/dist/Ionicons';
-import Blockie from '../../../../../components/Blockie';
 import Logo from '../../../../../components/Logo';
+import Blockie from '../../../../../components/scaffold-eth/Blockie';
 import useAccount from '../../../../../hooks/scaffold-eth/useAccount';
 import useNetwork from '../../../../../hooks/scaffold-eth/useNetwork';
 import { Account } from '../../../../../store/reducers/Accounts';
