+++ packages/reactnative/src/screens/Transfer/modules/Recipient.tsx
@@ -5,7 +5,7 @@ import { useModal } from 'react-native-modalfy';
 import { Text, TextInput } from 'react-native-paper';
 import { useSelector } from 'react-redux';
 import scaffoldConfig from '../../../../scaffold.config';
-import Blockie from '../../../components/Blockie';
+import Blockie from '../../../components/scaffold-eth/Blockie';
 import useAccount from '../../../hooks/scaffold-eth/useAccount';
 import { Account } from '../../../store/reducers/Accounts';
 import globalStyles from '../../../styles/globalStyles';
