+++ packages/reactnative/tsconfig.json
@@ -1,8 +1,3 @@
 {
   "extends": "@react-native/typescript-config/tsconfig.json",
-  "compilerOptions": {
-    "paths": {
-      "@/*": ["./*"]
-    }
-  }
 }
